# docs/swagger.yaml
basePath: /
definitions:
  genproto.AccountResponse:
    properties:
      account_id:
        type: string
      account_name:
        type: string
      account_type:
        type: string
      balance:
        type: number
      currency:
        type: string
      user_id:
        type: string
    type: object
  genproto.BudgetDeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  genproto.BudgetResponse:
    properties:
      amount:
        type: number
      budget_id:
        type: string
      category_id:
        type: string
      end_date:
        type: string
      period:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.CreateAccountRequest:
    properties:
      account_name:
        type: string
      balance:
        type: number
      currency:
        type: string
      id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.CreateAccountRes:
    properties:
      Message:
        type: string
    type: object
  genproto.CreateBudgetRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      end_date:
        type: string
      id:
        type: string
      period:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.DeleteResponse:
    properties:
      success:
        type: boolean
    type: object
...
  genproto.ListAccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/genproto.AccountResponse'
        type: array
    type: object
  genproto.ListBudgetsResponse:
    properties:
      budgets:
        items:
          $ref: '#/definitions/genproto.BudgetResponse'
        type: array
    type: object
...
  genproto.MessageResponsee:
    properties:
      Message:
        type: string
    type: object
...
  genproto.UpdateAccountRequest:
    properties:
      account_id:
        type: string
      account_name:
        type: string
      balance:
        type: number
      currency:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.UpdateBudgetRequest:
    properties:
      amount:
        type: number
      budget_id:
        type: string
      category_id:
        type: string
      end_date:
        type: string
      period:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Auth service API documentation
  title: api gat way
  version: "1.0"
paths:
  /account/create:
    post:
      consumes:
      - application/json
      description: Create a new account for a user
      parameters:
      - description: Create account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account created successfully
          schema:
            $ref: '#/definitions/genproto.CreateAccountRes'
        "500":
          description: Error while creating account
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Account
      tags:
      - Account
  /account/delete/{account_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing account by its ID
      parameters:
      - description: Account ID
        in: query
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            $ref: '#/definitions/genproto.DeleteResponse'
        "400":
          description: 'Missing required query parameter: account_id'
          schema:
            type: string
        "500":
          description: Error while deleting account
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Account
      tags:
      - Account
  /account/get/{account_id}:
    get:
      consumes:
      - application/json
      description: Retrieve an account's details by its ID
      parameters:
      - description: Account Id
        in: query
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account details
          schema:
            $ref: '#/definitions/genproto.AccountResponse'
        "400":
          description: 'Missing required query parameter: account_id'
          schema:
            type: string
        "500":
          description: Error while fetching account
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Account by Id
      tags:
      - Account
  /account/list:
    get:
      consumes:
      - application/json
      description: List all accounts for a user
      produces:
      - application/json
      responses:
        "200":
          description: List of accounts
          schema:
            $ref: '#/definitions/genproto.ListAccountsResponse'
        "500":
          description: Error while listing accounts
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Accounts
      tags:
      - Account
  /account/update:
    put:
      consumes:
      - application/json
      description: Update the details of an existing account
      parameters:
      - description: Update account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account updated successfully
          schema:
            $ref: '#/definitions/genproto.CreateAccountRes'
        "400":
          description: Invalid request or missing account_id
          schema:
            type: string
        "500":
          description: Error while updating account
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Account
      tags:
      - Account
  /budget/create:
    post:
      consumes:
      - application/json
      description: Create a new budget for a user
      parameters:
      - description: Create budget request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Budget created successfully
          schema:
            $ref: '#/definitions/genproto.MessageResponsee'
        "500":
          description: Error while creating budget
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Budget
      tags:
      - Budget
  /budget/delete/{budget_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing budget by its ID
      parameters:
      - description: Budget ID
        in: query
        name: budget_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget deleted successfully
          schema:
            $ref: '#/definitions/genproto.BudgetDeleteResponse'
        "400":
          description: 'Missing required query parameter: budget_id'
          schema:
            type: string
        "500":
          description: Error while deleting budget
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Budget
      tags:
      - Budget
  /budget/get/{budget_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a budget's details by its ID
      parameters:
      - description: Budget ID
        in: query
        name: budget_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget details
          schema:
            $ref: '#/definitions/genproto.BudgetResponse'
        "400":
          description: 'Missing required query parameter: budget_id'
          schema:
            type: string
        "500":
          description: Error while fetching budget
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Budget by ID
      tags:
      - Budget
  /budget/list:
    get:
      consumes:
      - application/json
      description: List all budgets for a user
      produces:
      - application/json
      responses:
        "200":
          description: List of budgets
          schema:
            $ref: '#/definitions/genproto.ListBudgetsResponse'
        "500":
          description: Error while listing budgets
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Budgets
      tags:
      - Budget
  /budget/update:
    put:
      consumes:
      - application/json
      description: Update the details of an existing budget
      parameters:
      - description: Update budget request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Budget updated successfully
          schema:
            $ref: '#/definitions/genproto.MessageResponsee'
        "400":
          description: Invalid request or missing budget_id
          schema:
            type: string
        "500":
          description: Error while updating budget
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Budget
      tags:
      - Budget
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"